+++ src/Modules/counter.js
@@ -1,6 +1,6 @@
 const counter = () => {
   const items = document.querySelectorAll('.meal');
-  return items.length
+  return items.length;
 };
 
 export default counter;
